start on page concat util
