+++ sedaro/src/sedaro/utils.py
@@ -8,6 +8,7 @@ import orjson
 from sedaro_base_client.api_client import ApiResponse
 from urllib3.response import HTTPResponse
 
+from .data_utils import update_metadata
 from .exceptions import SedaroApiException
 from .settings import BLOCKS, COMMON_API_KWARGS, STATUS
 
@@ -121,3 +122,31 @@ def get_class_from_module(module: ModuleType, target_class: str = None) -> type:
         raise AttributeError(err_msg)
 
     return filtered_classes[0][1]
+
+
+def concat_pages(pages: list[dict]) -> dict:
+    """Concatenates a list of Data Service pages into a single dictionary object.
+
+    Args:
+        pages (list[dict]): list of Data Service pages
+
+    Returns:
+        dict: concatenated data from all pages
+    """
+    result = {
+        'meta': {},
+        'series': {},
+        'stats': {},
+        'derived': {
+            'static': {},
+            'series': {},
+        }
+    }
+
+    first_page = False
+    for page in pages:
+        if first_page:
+            result['meta'] = page['meta']
+            first_page = False
+        else:
+            update_metadata(result['meta'], page['meta'])
